+++ tests/covalent_tests/executor/base_test.py
@@ -21,7 +21,13 @@
 """Tests for the Covalent executor base module."""
 
 import os
+import shutil
 import tempfile
+from pathlib import Path
+from unittest.mock import call
+
+import cloudpickle as pickle
+import pytest
 
 from covalent.executor import BaseExecutor
 
@@ -29,6 +35,31 @@ from covalent.executor import BaseExecutor
 class MockExecutor(BaseExecutor):
     def execute(self):
         pass
+
+
+class SubprocMock:
+    def __init__(self):
+        self.stdout = ""
+        self.stderr = ""
+
+
+def test_base_init(mocker):
+    mocker.patch.multiple(BaseExecutor, __abstractmethods__=set())
+
+    executor = BaseExecutor(
+        log_stdout="logout",
+        log_stderr="logerr",
+        conda_env="conda",
+        cache_dir="cache",
+        current_env_on_conda_fail=True,
+    )
+
+    assert executor.log_stdout == "logout"
+    assert executor.log_stderr == "logerr"
+    assert executor.conda_env == "conda"
+    assert executor.cache_dir == "cache"
+    assert executor.current_env_on_conda_fail
+    assert executor.current_env == ""
 
 
 def test_write_streams_to_file(mocker):
@@ -64,20 +95,99 @@ def test_write_streams_to_file(mocker):
         assert lines[0] == "absolute"
 
 
-def test_execute_in_conda_env(mocker):
+@pytest.mark.parametrize(
+    "conda_installed,successful_task,index", [(True, True, 0), (True, False, 1), (False, False, 2)]
+)
+def test_execute_in_conda_env(mocker, conda_installed, successful_task, index):
     """Test execute in conda enve method in Base Executor object."""
 
-    me = MockExecutor()
+    me = MockExecutor(cache_dir="cache_dir")
+    me.conda_path = "/path/to/conda"
 
-    mocker.patch("covalent.executor.BaseExecutor.get_conda_path", return_value=False)
-    conda_env_fail_mock = mocker.patch("covalent.executor.BaseExecutor._on_conda_env_fail")
-    me.execute_in_conda_env(
-        "function",
-        "function_version",
-        "args",
-        "kwargs",
-        "conda_env",
-        "cache_dir",
-        "node_id",
+    subproc = SubprocMock()
+    subproc.stdout = ""
+    subproc.stderr = "Error"
+    subproc.returncode = 0 if successful_task else 1
+
+    def test_func(a, /, *, b):
+        return a + b
+
+    result_filename = f"cache_dir/result_pickle_file{index}.pkl"
+    with open(result_filename, "wb") as f:
+        pickle.dump(3, f)
+
+    def get_temp_file(dir, delete=None, mode=None):
+        if delete is not None:
+            return open(f"pickle_file{index}.pkl", "wb")
+        elif mode == "w":
+            return open(f"script_file{index}.sh", "w")
+
+    get_conda_path_mock = mocker.patch(
+        "covalent.executor.BaseExecutor.get_conda_path", return_value=conda_installed
     )
-    conda_env_fail_mock.assert_called_once_with("function", "args", "kwargs", "node_id")
+    conda_env_fail_mock = mocker.patch("covalent.executor.base.BaseExecutor._on_conda_env_fail")
+    pickle_dump_mock = mocker.patch("covalent.executor.base.pickle.dump", return_value=None)
+    pickle_load_mock = mocker.patch("covalent.executor.base.pickle.load", return_value=3)
+    tempfile_mock = mocker.patch("tempfile.NamedTemporaryFile", side_effect=get_temp_file)
+    path_exists_mock = mocker.patch("os.path.exists", return_value=conda_installed)
+    subprocess_mock = mocker.patch("subprocess.run", return_value=subproc)
+
+    result = me.execute_in_conda_env(
+        fn=test_func,
+        fn_version="function_version",
+        args=[1],
+        kwargs={"b": 2},
+        conda_env="conda_env",
+        cache_dir="cache_dir",
+        node_id=100,
+    )
+
+    get_conda_path_mock.assert_called_once_with()
+    if (not conda_installed) or (not successful_task):
+        conda_env_fail_mock.assert_called_once_with(test_func, [1], {"b": 2}, 100)
+        if conda_installed:
+            os.remove(f"pickle_file{index}.pkl")
+            os.remove(f"script_file{index}.sh")
+        return
+
+    assert Path(f"pickle_file{index}.pkl").exists()
+    assert Path(f"script_file{index}.sh").exists()
+
+    tempfile_mock.assert_has_calls(
+        [call(dir="cache_dir", delete=False), call(dir="cache_dir", mode="w")]
+    )
+    pickle_dump_mock.assert_called_once()
+    path_exists_mock.assert_called_once_with("/path/to/../etc/profile.d/conda.sh")
+    subprocess_mock.assert_called_once_with(
+        ["bash", f"script_file{index}.sh"], capture_output=True, encoding="utf-8"
+    )
+    pickle_load_mock.assert_called_once()
+
+    os.remove(f"pickle_file{index}.pkl")
+    os.remove(f"script_file{index}.sh")
+
+
+def test_conda_env_fail(mocker):
+    pass
+
+
+def test_get_conda_envs(mocker):
+    conda_env_str = """\
+# conda environments:
+#
+base * /home/user/miniconda3
+test /home/user/miniconda3/envs/test\
+"""
+
+    subproc = SubprocMock()
+    subproc.stdout = conda_env_str
+    subprocess_mocker = mocker.patch("subprocess.run", return_value=subproc)
+
+    me = MockExecutor()
+    me.get_conda_envs()
+
+    subprocess_mocker.assert_called_once_with(
+        ["conda", "env", "list"], capture_output=True, encoding="utf-8"
+    )
+    assert me.conda_envs == ["base", "test"]
+    assert me.current_env == "base"