+++ CHANGELOG.md
@@ -5,11 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.34.2] - 2022-03-25
+## [0.34.4] - 2022-03-28
 
 ### Fixed 
 
 - Moved `example_dispatch.py` into `tests/` directory.
+
+## [0.34.3] - 2022-03-27
+
+### Added
+
+- Tests for lattice.py
+
+## [0.34.2] - 2022-03-27
+
+### Added
+
+- Unit tests for the base executor, the results manager, the logger, and leptons
 
 ## [0.34.1] - 2022-03-24